Part 1: Social network management and SNA
A social network (facebook, twitter, etc.) consists of nodes and edges. In such a network,
nodes refer to individuals or things within the network while edges or ties refer to relations
or interaction.
Wetherell et al. [1] describes Social Network Analysis (SNA), also called as structural analysis,
as \(1) conceptualizes social structure as a network with ties connecting members and chan-
nelling resources, (2) focuses on the characteristics of ties rather than on the characteristics
of the individual members, and (3) views communities as 'personal communities', that is, as
networks of individual relations that people foster, maintain, and use in the course of their
daily lives". In brief, SNA is the process of analyzing the social structure using network or
graph theories.
To accomplish this part of the assignment, it is expected from you to manage and investigate
the relation among the nodes in the network, demonstrated in Fig. 1, using File (I/O),
dictionaries, lists and sets.
At the beginning phase of this part, your implementation should rst read the le \com-
mandsP1.txt" which will be shared then produce in the direction of those commands. The
implementation should have those functionalities over the network (Fig. 1), each of them is
introduced below:

1. Add user: In this option, user gives a name to the node and ties it to an existing node.
Note that, user is not allowed to give an existing name to a newly created node nor to
tie it to non-existing node. This process requires two inputs, rst indicates new node
and latter indicates existing node. Command type should be as follows:
AU < newuser > < existinguser >
2. Remove user: In this option, a node whose name is dened by user and its all relations
are removed. User must be restricted to remove a non-existing node. This process
requires only one input which indicates existing node. Command type should be as
follows:
RU < username >
3. Add new relation: A new relation between two nodes is created in this option. Note
that, the criteria for a new relation are that (1)there must not exist such a relation
created before and (2)both nodes must exist in the network. This process requires two
inputs, source node and target node. Command type should be as follows:
AR < sourceuser >< targetuser >
4. Remove existing relation: Within the scope of this option, user determines the
nodes' names and must be avoided when he/she attempts to remove a non-existing
relation or inputs a non-existing node name. As in the previous process, here two
inputs are required, source node and target node. Command type is given below:
RR < sourceuser >< targetuser >
5. Rank users: User performs degree (popularity) analysis for each node and displays
top n users depending on their popularities in this option. Here n is determined by user
and must not be greater than total number of user in the network. This process takes
only one input \n". Command type should be as follows:
PA < toplistsize >
6. Suggest friendship: In the last option, user determines Mutuality Degree (MD) and
a node \i" to list its possible friends. Here, MD is a threshold identifying the one
as suggested friend for i. For example, user attempts to analyze friendship status to
suggest friends for node C when MD = 3. As seen in Fig. 2, 'C' has 3 friends f'E',
'O', 'R'g and each of them has 5 f'C', 'I', 'J', 'L', Pg, 6 f'A', 'C', 'G', 'M', 'P', 'T'g,
and 9 f'B', 'C', 'D', 'F', 'H', 'K', 'L', 'N', 'P'g friends, respectively. When considering
to the friend lists of C's friends, it can be noticed that only node P reaches or exceeds
MD, since C's at least 3 friends are friend with P. For MD = 2, the suggestion list will
become P and L. Note that, MD must be greater than 1 and user cannot determine a
non-existing node or a node having friend less than MD.
SA < username >< MD >