Part 2: Possible path nder
This part of the assignment is based on this scenario: Assume a traveler \t" plans to visit some
cities in the network given in Fig. 3. As seen in the gure, there is one-way transportation
between any of two cities (directions are represented by arrows). Starting from a city, say x, t
can only visit the cities whose maximum distance to x is 3. For instance, if t starts from 'C'
then the reachable city list is f'E', 'G', 'H', 'A', 'K', 'N', 'O', 'M', 'I'g; for x='J' reachable
city list is f'K', 'N', 'O'g.
As in the previous part, your implementation should rst read the le, (commandsP2.txt),
consisting of starting cities then lists reachable cities as given in Fig. 4.
The implementation investigates possible cities based on the descriptions above and it displays
the reachable cities as shown in Fig. 4 for the cities given in the input file.